+++ tests/conftest.py
@@ -10,12 +10,7 @@ from fastapi.testclient import TestClient
 from neo4j import Driver, Session, SummaryCounters
 from pytest import MonkeyPatch
 
-from mex.backend.auxiliary.ldap import (
-    extracted_organizational_unit,
-    extracted_primary_source_ldap,
-)
-from mex.backend.auxiliary.primary_source import extracted_primary_source_orcid
-from mex.backend.auxiliary.wikidata import extracted_primary_source_wikidata
+from mex.artificial.helpers import generate_artificial_extracted_items
 from mex.backend.graph.connector import GraphConnector
 from mex.backend.identity.provider import GraphIdentityProvider
 from mex.backend.main import app
@@ -24,6 +19,7 @@ from mex.backend.settings import BackendSettings
 from mex.backend.types import APIKeyDatabase, APIUserDatabase
 from mex.common.connector import CONNECTOR_STORE
 from mex.common.models import (
+    EXTRACTED_MODEL_CLASSES_BY_NAME,
     MEX_PRIMARY_SOURCE_STABLE_TARGET_ID,
     AdditiveOrganizationalUnit,
     AnyExtractedModel,
@@ -321,6 +317,17 @@ def dummy_data(
     }
 
 
+@pytest.fixture
+def artificial_extracted_items() -> list[AnyExtractedModel]:
+    return generate_artificial_extracted_items(
+        locale="de_DE",
+        seed=42,
+        count=25,
+        chattiness=16,
+        stem_types=EXTRACTED_MODEL_CLASSES_BY_NAME,
+    )
+
+
 def _match_organization_items(dummy_data: dict[str, AnyExtractedModel]) -> None:
     # TODO(ND): replace this crude item matching implementation (stopgap MX-1530)
     connector = GraphConnector.get()
@@ -347,9 +354,19 @@ def load_dummy_data(
     dummy_data: dict[str, AnyExtractedModel],
 ) -> dict[str, AnyExtractedModel]:
     """Ingest dummy data into the graph."""
-    GraphConnector.get().ingest(list(dummy_data.values()))
+    connector = GraphConnector.get()
+    connector.ingest(list(dummy_data.values()))
     _match_organization_items(dummy_data)
     return dummy_data
+
+
+@pytest.fixture
+def load_artificial_extracted_items(
+    artificial_extracted_items: list[AnyExtractedModel],
+) -> list[AnyExtractedModel]:
+    """Ingest artificial data into the graph."""
+    GraphConnector.get().ingest(artificial_extracted_items)
+    return artificial_extracted_items
 
 
 @pytest.fixture
@@ -384,7 +401,8 @@ def load_dummy_rule_set(
     organizational_unit_rule_set_request: OrganizationalUnitRuleSetRequest,
     load_dummy_data: dict[str, AnyExtractedModel],
 ) -> OrganizationalUnitRuleSetResponse:
-    GraphConnector.get().ingest(
+    connector = GraphConnector.get()
+    connector.ingest(
         [
             load_dummy_data["primary_source_2"],
             load_dummy_data["organizational_unit_1"],
@@ -398,12 +416,3 @@ def load_dummy_rule_set(
             stable_target_id=load_dummy_data["organizational_unit_2"].stableTargetId,
         ),
     )
-
-
-@pytest.fixture(autouse=True)
-def reset_caches() -> None:
-    """Reset the caches for each test."""
-    extracted_primary_source_wikidata.cache_clear()
-    extracted_primary_source_ldap.cache_clear()
-    extracted_primary_source_orcid.cache_clear()
-    extracted_organizational_unit.cache_clear()