+++ tests/graph/test_query.py
@@ -176,6 +176,7 @@ def test_fetch_extracted_or_rule_items(
     query = query_builder.fetch_extracted_or_rule_items(
         filter_by_query_string=filter_by_query_string,
         filter_by_stable_target_id=filter_by_stable_target_id,
+        filter_by_reference_to_merged_item=False,
     )
     assert str(query) == expected
 
@@ -203,7 +204,7 @@ def test_fetch_extracted_or_rule_items(
         RETURN extracted_or_rule_node, merged_node
     }
     WITH DISTINCT merged_node AS merged_node
-    MATCH (merged_node)<-[:stableTargetId]-(:ExtractedThis|ExtractedThat|ExtractedOther)-[:hadPrimarySource]->(referenced_merged_node_to_filter_by)
+    MATCH (merged_node)<-[:stableTargetId]-()-[:hadPrimarySource]->(referenced_merged_node_to_filter_by)
     WHERE
         ANY(label IN labels(merged_node) WHERE label IN $labels)
         AND merged_node.identifier = $identifier
@@ -221,7 +222,7 @@ CALL () {
         RETURN extracted_or_rule_node, merged_node
     }
     WITH DISTINCT merged_node AS merged_node
-    MATCH (merged_node)<-[:stableTargetId]-(:ExtractedThis|ExtractedThat|ExtractedOther)-[:hadPrimarySource]->(referenced_merged_node_to_filter_by)
+    MATCH (merged_node)<-[:stableTargetId]-()-[:hadPrimarySource]->(referenced_merged_node_to_filter_by)
     WHERE
         ANY(label IN labels(merged_node) WHERE label IN $labels)
         AND merged_node.identifier = $identifier