+++ tests/ingest/test_main.py
@@ -8,11 +8,11 @@ from pytest import MonkeyPatch
 from starlette import status
 
 from mex.backend.graph.connector import GraphConnector
+from mex.backend.merged.helpers import search_merged_items_in_graph
 from mex.common.models import (
     EXTRACTED_MODEL_CLASSES,
     RULE_SET_RESPONSE_CLASSES,
     AnyExtractedModel,
-    ItemsContainer,
     PaginatedItemsContainer,
 )
 
@@ -20,11 +20,11 @@ Payload = dict[str, list[dict[str, Any]]]
 
 
 @pytest.fixture
-def post_payload(dummy_data: dict[str, AnyExtractedModel]) -> Payload:
+def post_payload(artificial_extracted_items: list[AnyExtractedModel]) -> Payload:
     payload = defaultdict(list)
-    for model in dummy_data.values():
+    for model in artificial_extracted_items:
         payload["items"].append(model.model_dump())
-    return cast(Payload, dict(payload))
+    return cast("Payload", dict(payload))
 
 
 @pytest.mark.integration
@@ -32,34 +32,47 @@ def test_bulk_insert_empty(client_with_api_key_write_permission: TestClient) -> 
     response = client_with_api_key_write_permission.post(
         "/v0/ingest", json={"items": []}
     )
-    assert response.status_code == status.HTTP_201_CREATED, response.text
-    assert response.json() == {"items": []}
+    assert response.status_code == status.HTTP_204_NO_CONTENT, response.text
+    assert response.text == ""
 
 
 @pytest.mark.integration
 def test_bulk_insert(
     client_with_api_key_write_permission: TestClient,
     post_payload: Payload,
-    dummy_data: dict[str, AnyExtractedModel],
+    artificial_extracted_items: list[AnyExtractedModel],
 ) -> None:
-    # post the dummy data to the ingest endpoint
+    # post the artificial data to the ingest endpoint
     response = client_with_api_key_write_permission.post(
         "/v0/ingest", json=post_payload
     )
 
-    # assert the response are the dummy data items
-    assert response.status_code == status.HTTP_201_CREATED, response.text
-    assert ItemsContainer[AnyExtractedModel].model_validate(
-        response.json()
-    ).items == list(dummy_data.values())
+    # assert the response are the artificial data items
+    assert response.status_code == status.HTTP_204_NO_CONTENT, response.text
+    assert response.text == ""
 
     # verify the nodes have actually been stored in the database
-    graph = GraphConnector.get()
-    result = graph.fetch_extracted_items(None, None, None, 1, len(dummy_data))
+    connector = GraphConnector.get()
+    result = connector.fetch_extracted_items(
+        None, None, None, None, 1, len(artificial_extracted_items)
+    )
     result_container = PaginatedItemsContainer[AnyExtractedModel].model_validate(
         result.one()
     )
-    assert set(result_container.items) == set(dummy_data.values())
+    assert set(result_container.items) == set(artificial_extracted_items)
+
+    # verify that the merging worked correctly
+    result_merged = search_merged_items_in_graph(
+        None, None, None, None, 1, len(artificial_extracted_items)
+    )
+
+    assert (
+        {merged_item.identifier for merged_item in result_merged.items}
+        == {
+            artifical_item.stableTargetId
+            for artifical_item in artificial_extracted_items
+        }
+    )  # compare length (indirectly) and if IDs of extracted & merged items are identical
 
 
 def test_bulk_insert_malformed(
@@ -96,7 +109,6 @@ def test_bulk_insert_malformed(
 def test_bulk_insert_mocked(
     client_with_api_key_write_permission: TestClient,
     post_payload: Payload,
-    dummy_data: dict[str, AnyExtractedModel],
     monkeypatch: MonkeyPatch,
 ) -> None:
     monkeypatch.setattr(GraphConnector, "_merge_item", MagicMock())
@@ -104,7 +116,4 @@ def test_bulk_insert_mocked(
     response = client_with_api_key_write_permission.post(
         "/v0/ingest", json=post_payload
     )
-    assert response.status_code == status.HTTP_201_CREATED, response.text
-    assert ItemsContainer[AnyExtractedModel].model_validate(
-        response.json()
-    ).items == list(dummy_data.values())
+    assert response.status_code == status.HTTP_204_NO_CONTENT, response.text