+++ CHANGELOG.md
@@ -19,6 +19,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [0.34.0] - 2025-04-29
+
+### Added
+
+- added `hadPrimarySource` parameter to extracted items search
+- added `/v0/_system/metrics` endpoint for prometheus metrics
+- add artificial data creator helper and use in test
+
+### Changes
+
+- change unit and primary source helpers from caching to graph lookup
+
+### Removed
+
+- BREAKING: remove response body from ingest endpoint
+
 ## [0.33.2] - 2025-04-16
 
 ### Removed